+++ packages/shared/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@iavofficial/frontend-framework-shared-v1.0.0-dev.9](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.8...@iavofficial/frontend-framework-shared-v1.0.0-dev.9) (2025-11-21)
+
+### Bug Fixes
+
+* Fixed imports for defaultContentSelectionElement. Fixed types regarding cyclic dependencies. ([738f3a6](https://github.com/iavofficial/IAVFrontendFramework/commit/738f3a65b650d5d4a2185a70168de63ef0b7808f))
+
+## [@iavofficial/frontend-framework-shared-v1.0.0-dev.8](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.7...@iavofficial/frontend-framework-shared-v1.0.0-dev.8) (2025-11-21)
+
+### Bug Fixes
+
+* Fixed circular dependency ([2a2fd8e](https://github.com/iavofficial/IAVFrontendFramework/commit/2a2fd8e09439a995b3c7a5f16f651961d4b55308))
+
 ## [@iavofficial/frontend-framework-shared-v1.0.0-dev.7](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.6...@iavofficial/frontend-framework-shared-v1.0.0-dev.7) (2025-11-21)
 
 ### Bug Fixes